restructure Flask class in main.py, put calculation logic in a separate class loanCalculator(loan_calculator.py)
